Proj/pyproject.toml
[project]
name = "proj"
version = "0.1.0"
description = "阿里云通义千问 API 调用示例（命令行 + Gradio Web UI）"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "gradio>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "pytest>=8.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
